requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg[binary]==3.1.18
django-environ==0.11.2
argon2-cffi==23.1.0
cryptography==42.0.5
stripe==8.11.0
coinbase-commerce==1.1.1
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.3
django-ratelimit==4.1.0
django-idempotency-key==1.1.1
pytest-django==4.8.0
factory-boy==3.3.0
ruff==0.3.5
black==24.4.0
gunicorn==22.0.0
whitenoise==6.6.0
django-geoip2==3.1
